+++ admin/inertia/layouts/SettingsLayout.tsx
@@ -1,5 +1,6 @@
 import {
   IconArrowBigUpLines,
+  IconBook,
   IconBuildingStore,
   IconCarrot,
   IconChartBar,
@@ -45,6 +46,7 @@ export default function SettingsLayout({ children }: { children: React.ReactNode
       current: false,
     },
     { name: 'System', href: '/settings/system', icon: IconSettings, current: false },
+    { name: 'Documentation', href: '/docs', icon: IconBook, current: false },
     { name: 'Support the Project', href: '/settings/support', icon: IconHeart, current: false },
     { name: 'Legal Notices', href: '/settings/legal', icon: IconGavel, current: false },
   ]